requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
openai==1.30.5
requests==2.32.3
pyyaml==6.0.1
